# YoFi.Experiments.WebApi/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "YoFi.Experiments.WebApi, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/Reports": {
      "get": {
        "tags": [
          "Reports"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReportDefinition"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Reports"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportParameters"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportParameters"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReportParameters"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WireReport"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/Reports/Summary": {
      "post": {
        "tags": [
          "Reports"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportParameters"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportParameters"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReportParameters"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/WireReport"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/Transactions": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "operationId": "ListTransactions",
        "parameters": [
          {
            "name": "Query",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Order",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "View",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "All",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionWireQueryResult"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Transactions"
        ],
        "operationId": "CreateTransaction",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Transaction"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Transaction"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Transaction"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          }
        }
      }
    },
    "/Transactions/{id}": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "operationId": "GetTransaction",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Transactions"
        ],
        "operationId": "DeleteTransaction",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Transactions"
        ],
        "operationId": "UpdateTransaction",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Transaction"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Transaction"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Transaction"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/Transactions/Download/{year}": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "operationId": "DownloadTransactions",
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "allyears",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "query",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {
                "schema": {
                  "$ref": "#/components/schemas/FileStream"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileStream"
                }
              }
            }
          }
        }
      }
    },
    "/WeatherForecast": {
      "get": {
        "tags": [
          "WeatherForecast"
        ],
        "operationId": "GetWeatherForecast",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecast"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecast"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecast"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ColumnLabel": {
        "type": "object",
        "properties": {
          "uniqueID": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "isTotal": {
            "type": "boolean"
          },
          "isSortingAfterTotal": {
            "type": "boolean"
          },
          "isCalculated": {
            "type": "boolean"
          },
          "displayAsPercent": {
            "type": "boolean"
          },
          "leafNodesOnly": {
            "type": "boolean"
          },
          "isSeries": {
            "type": "boolean"
          },
          "priority": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "FileStream": {
        "type": "object",
        "properties": {
          "canTimeout": {
            "type": "boolean",
            "readOnly": true
          },
          "readTimeout": {
            "type": "integer",
            "format": "int32"
          },
          "writeTimeout": {
            "type": "integer",
            "format": "int32"
          },
          "handle": {
            "$ref": "#/components/schemas/IntPtr"
          },
          "canRead": {
            "type": "boolean",
            "readOnly": true
          },
          "canWrite": {
            "type": "boolean",
            "readOnly": true
          },
          "safeFileHandle": {
            "$ref": "#/components/schemas/SafeFileHandle"
          },
          "name": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "isAsync": {
            "type": "boolean",
            "readOnly": true
          },
          "length": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          },
          "position": {
            "type": "integer",
            "format": "int64"
          },
          "canSeek": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "IWirePageInfo": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "pageSize": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "firstItem": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "numItems": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "totalPages": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "totalItems": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "IWireQueryParameters": {
        "type": "object",
        "properties": {
          "query": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "page": {
            "type": "integer",
            "format": "int32",
            "nullable": true,
            "readOnly": true
          },
          "order": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "view": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "all": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "IntPtr": {
        "type": "object",
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "ReportDefinition": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "nullable": true
          },
          "source": {
            "type": "string",
            "nullable": true
          },
          "sourceParameters": {
            "type": "string",
            "nullable": true
          },
          "customColumns": {
            "type": "string",
            "nullable": true
          },
          "wholeYear": {
            "type": "boolean",
            "nullable": true
          },
          "yearProgress": {
            "type": "boolean",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "sortOrder": {
            "type": "string",
            "nullable": true
          },
          "withMonthColumns": {
            "type": "boolean",
            "nullable": true
          },
          "withTotalColumn": {
            "type": "boolean",
            "nullable": true
          },
          "skipLevels": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "numLevels": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "displayLevelAdjustment": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ReportParameters": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "nullable": true
          },
          "year": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "month": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "showmonths": {
            "type": "boolean",
            "nullable": true
          },
          "level": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RowLabel": {
        "type": "object",
        "properties": {
          "uniqueID": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "isTotal": {
            "type": "boolean"
          },
          "isSortingAfterTotal": {
            "type": "boolean"
          },
          "parent": {
            "$ref": "#/components/schemas/RowLabel"
          },
          "level": {
            "type": "integer",
            "format": "int32"
          },
          "collector": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SafeFileHandle": {
        "type": "object",
        "properties": {
          "isClosed": {
            "type": "boolean",
            "readOnly": true
          },
          "isInvalid": {
            "type": "boolean",
            "readOnly": true
          },
          "isAsync": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "Split": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "category": {
            "type": "string",
            "nullable": true
          },
          "memo": {
            "type": "string",
            "nullable": true
          },
          "transactionID": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "Transaction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "timestamp": {
            "type": "string",
            "format": "date"
          },
          "payee": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "category": {
            "type": "string",
            "nullable": true
          },
          "memo": {
            "type": "string",
            "nullable": true
          },
          "bankReference": {
            "type": "string",
            "nullable": true
          },
          "hidden": {
            "type": "boolean",
            "nullable": true
          },
          "imported": {
            "type": "boolean",
            "nullable": true
          },
          "selected": {
            "type": "boolean",
            "nullable": true
          },
          "receiptUrl": {
            "type": "string",
            "nullable": true
          },
          "splits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Split"
            },
            "nullable": true
          },
          "hasSplits": {
            "type": "boolean",
            "readOnly": true
          },
          "isSplitsOK": {
            "type": "boolean",
            "readOnly": true
          },
          "strippedPayee": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "TransactionWireQueryResult": {
        "type": "object",
        "properties": {
          "parameters": {
            "$ref": "#/components/schemas/IWireQueryParameters"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Transaction"
            },
            "nullable": true
          },
          "pageInfo": {
            "$ref": "#/components/schemas/IWirePageInfo"
          }
        },
        "additionalProperties": false
      },
      "WeatherForecast": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "temperatureC": {
            "type": "integer",
            "format": "int32"
          },
          "temperatureF": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "summary": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "WireReport": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "definition": {
            "type": "string",
            "nullable": true
          },
          "columnLabels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ColumnLabel"
            },
            "nullable": true
          },
          "rowLabels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RowLabel"
            },
            "nullable": true
          },
          "displayLevelAdjustment": {
            "type": "integer",
            "format": "int32"
          },
          "grandTotal": {
            "type": "number",
            "format": "double"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WireReportLine"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "WireReportLine": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "values": {
            "type": "array",
            "items": {
              "type": "number",
              "format": "double"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}